continue splitting large modules
